README
We can modify our outcomes by setting either lowercase, uppercase, digits or punctuation to False. You can do that by typing "lowercase" in the terminal in order to NOT have lowercase letters in your password likewise we can do that for "uppercase", "digits" and "punctuation". NOTE that if you set all attributes to False the application with exit.